requirements.txt
matplotlib~=3.5.2
scikit-image
torch~=1.8.1
torchvision~=0.9.1
numpy~=1.21.6
pillow~=9.1.1
tqdm~=4.64.0
jupyter==1.0.0
opencv-python==4.6.0.66
scikit-learn==1.0.2
pandas==1.3.5
